+++ src/dialogs/dialogs-common.ts
@@ -1,4 +1,4 @@
-import { Color, Enums, Font, ImageSource, View, ViewBase } from '@nativescript/core';
+import { Color, CoreTypes, Font, ImageSource, View, ViewBase } from '@nativescript/core';
 
 export interface MDCAlertControlerOptions {
     buttonFont?: Font;