+++ typedoc.js
@@ -3,11 +3,8 @@ module.exports = {
     out: 'docs',
     exclude: ['**/node_modules/**', '**/*.spec.ts', '*typings*'],
     name: 'Nativescript UI Material Components',
-    // excludePrivate: true,
-    // ignoreCompilerErrors: true,
-    // excludeNotExported: true,
-    // includeDeclarations: true,
     excludePrivate: true,
+    excludeProtected: true,
     excludeExternals: true,
     tsconfig: 'tsconfig.doc.json',
     readme: 'README.md'